esp32-s3: move render function into new module which should result into embedded crate
